[Code Gen] Add code generation for functions.
Nodes get an "output_computation" member for generating the code of
whatever computation they represent.

Definitions change from having a Node and some arguments as their
object to having a really vague 'Defineable' that can be either a
function or a type. The type can have several constructors.

Function and Type are derived from Defineable. Type is not fully
fledged. Function is a Node and some arguments. Defineables have
prototypes, info tables and definitions that can be output by virtual
functions. The version of these for Function will use the Node's
output_computation member to generate the body of the function.