[Boxing] A tiny step towards unboxed data.
The tokenization of the 'special' lexemes has been slightly refactored.

The tokenizations of the | PIPE and -> ARROW_TO operators used in case
expressions have been upgraded.

The hash # character can be second or later in a varID or ConID as part
of starting on unboxed data.

prelude-ctor.x1 has had an Int datatype and 'plus' function added which
are written as though Int was a box with an MkInt# constructor and plus
used case expressions based on that constructor to deconstruct and add
the terms and reconstruct the result.

base.c has had a *# operator function added which plays a part in the
implementation of the plus function in prelude-ctor.x1