[Awkward squad] ccall primitive added
Modified parse_primary to potentially recognize a ccall form as a
call to a c function with strictness of argument evaluation of course.
ccall has to be 'maintained' in order to correctly call a given
function. It won't work with arbitrary functions.

The potential value of this is, at present, locked away in some test
programs that use it to call putchar.

Currently this primitive does NOT return a World like the other guy
does.