[Awkward squad] Test programs for a certain style of IO
Test programs that use ccall putchar to achieve output. A mysterious
world parameter is involved.

The idea would be that the world would not be mentioned explicitly in
code, but when something like this is someday baked in to the RTS,
base.c would pass the parameter to whatever x1's main returns.

By the way, io3.x1's printnum is a little bogus, it prints the number
backwards. This is why shownum's implementation in the prelude has a
reverse within.

Next stop: Input.
Next stop: type-checking. Actually had crashes because intermediate
versions were not correct as to types.